LayoutTests/imported/w3c/web-platform-tests/css/cssom/cssstyledeclaration-setter-order-expected.txt


FAIL setProperty with existing longhand should change order assert_array_equals: Property order should match after setting existing property expected property 0 to be "bottom" but got "top" (expected array ["bottom", "left", "right", "top"] got ["top", "bottom", "left", "right"])
FAIL invoke property setter with existing longhand should change order assert_array_equals: Property order should match after setting existing property expected property 0 to be "bottom" but got "top" (expected array ["bottom", "left", "right", "top"] got ["top", "bottom", "left", "right"])
FAIL setProperty with existing shorthand should change order assert_array_equals: Property order should match after setting an existing shorthand expected property 0 to be "top" but got "margin-top" (expected array ["top", "padding-top", "padding-right", "padding-bottom", "padding-left", "margin-top", "margin-right", "margin-bottom", "margin-left"] got ["margin-top", "margin-right", "margin-bottom", "margin-left", "top", "padding-top", "padding-right", "padding-bottom", "padding-left"])
FAIL invoke property setter with existing shorthand should change order assert_array_equals: Property order should match after setting an existing shorthand expected property 0 to be "top" but got "margin-top" (expected array ["top", "padding-top", "padding-right", "padding-bottom", "padding-left", "margin-top", "margin-right", "margin-bottom", "margin-left"] got ["margin-top", "margin-right", "margin-bottom", "margin-left", "top", "padding-top", "padding-right", "padding-bottom", "padding-left"])
